requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
openai
email-validator
google-generativeai
bcrypt==4.0.1